// src/infra/api/model/editarEstabelecimentoDto.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Minha API
 * Documentação oficial da API
 * OpenAPI spec version: 1.0.0
 */

export interface EditarEstabelecimentoDto {
  /** @maxLength 255 */
  nome: string;
  descricao?: string;
  sobreNos?: string;
  /** @maxLength 20 */
  telefone?: string;
  /** @maxLength 20 */
  priceRange?: string;
  /** @maxLength 20 */
  status?: string;
  verificado?: boolean;
  destaque?: boolean;
  logoUrl?: string;
  coverUrl?: string;
  /** @maxLength 255 */
  slug?: string;
}