// src/infra/api/model/estabelecimentoModelOutput.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Minha API
 * Documentação oficial da API
 * OpenAPI spec version: 1.0.0
 */

export interface EstabelecimentoModelOutput {
  id?: number;
  /**
   * @minLength 1
   * @maxLength 255
   */
  nome: string;
  /**
   * @minLength 1
   * @maxLength 255
   * @pattern ^[a-z0-9-]+$
   */
  slug: string;
  /**
   * @maxLength 5000
   * @nullable
   */
  descricao?: string | null;
  /** @nullable */
  sobreNos?: string | null;
  /**
   * @maxLength 20
   * @nullable
   */
  telefone?: string | null;
  /** @nullable */
  priceRange?: string | null;
  /** @nullable */
  status?: string | null;
  verificado: boolean;
  destaque: boolean;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  /**
   * @minimum 0
   * @maximum 5
   */
  ratingAvg: number;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  ratingCount: number;
  /**
   * @maxLength 255
   * @nullable
   */
  googlePlaceId?: string | null;
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  deletedAt?: string | null;
}